refactor: name modal + waiting panel
